Put MFIntegrator and PAIntegrator in the same file.
